+++ apiphany-core/src/main/java/org/apiphany/lang/Lifecycle.java
@@ -2,20 +2,24 @@ package org.apiphany.lang;
 
 /**
  * Enum for life-cycle management of objects.
+ * <p>
+ * This enum defines ownership responsibility, not resource lifetime and is defined at the point of the
+ * {@link ScopedResource} creation. It indicates whether the resource is managed (owned) by the current scope or
+ * unmanaged (owned by an external container or framework).
  *
  * @author Radu Sebastian LAZIN
  */
 public enum Lifecycle {
 
 	/**
-	 * Managed life-cycle, meaning that the object is managed by a container or framework, and its life-cycle is controlled
-	 * by that container or framework.
+	 * Managed life-cycle, meaning that the object is owned by the current scope, and its life-cycle is controlled by
+	 * whoever owns the {@link ScopedResource}.
 	 */
 	MANAGED,
 
 	/**
-	 * Unmanaged life-cycle, meaning that the object is not managed by any container or framework, and its life-cycle is
-	 * controlled by the application code.
+	 * Unmanaged life-cycle, meaning that the object is managed by a container or framework, and its life-cycle is
+	 * controlled by the container or framework not by the owner of the {@link ScopedResource}.
 	 */
 	UNMANAGED;